Change visual behaviour slightly
Axes now display their minimum value as well.

Decreased top and increased bottom margin for diagrams.

Added axis-like value display elements to color scales.
Added left and right margins for color scales.

Overrode paintBorder() methods of visual elements to keep border
thickness updates more reliably.

Now once the HoverLabel exits a Hoverable component, its caption is
erased.

+++ Implementation Phase/GELF/src/main/java/gelf/view/diagrams/components/DiagramLine.java
@@ -38,7 +38,7 @@ public abstract class DiagramLine extends DiagramComponent {
 		return this.start;
 	}
 
-	public void setStartInFrame(float x1, float y1) {
+	public void setStartInFrame(double x1, double y1) {
 		this.start.setXPos(x1);
 		this.start.setYPos(y1);
 		
@@ -49,7 +49,7 @@ public abstract class DiagramLine extends DiagramComponent {
 		return this.end;
 	}
 
-	public void setEndInFrame(float x2, float y2) {
+	public void setEndInFrame(double x2, double y2) {
 		this.end.setXPos(x2);
 		this.end.setYPos(y2);
